requirements.txt
altair==5.1.2
attrs==23.1.0
blinker==1.6.3
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==1.1.0
fiona==1.9.5
folium==0.14.0
fonttools==4.43.1
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaleido==0.2.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
numerize==0.12
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
patsy==0.5.3
Pillow==10.0.1
plotly==5.17.0
plotly-express==0.4.1
protobuf==4.24.4
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
statsmodels==0.14.0
streamlit==1.27.2
streamlit-carousel==0.0.4
streamlit-folium==0.15.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
wordcloud==1.9.2
XlsxWriter==3.1.9
zipp==3.17.0
